apps/python_client/pyproject.toml
[tool.poetry]
name = "python-client"
version = "0.1.0"
description = ""
authors = ["HEKUCHAN <heitorhirose@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.70.0"
grpcio-tools = "^1.70.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
